Isolate phenomodels inside their own blueprint
