+++ scout/server/app.py
@@ -24,6 +24,7 @@ from .blueprints import (
     login,
     managed_variants,
     panels,
+    phenomodels,
     phenotypes,
     public,
     variant,
@@ -183,6 +184,7 @@ def register_blueprints(app):
     app.register_blueprint(api.api_bp)
     app.register_blueprint(alignviewers.alignviewers_bp)
     app.register_blueprint(phenotypes.hpo_bp)
+    app.register_blueprint(phenomodels.phenomodels_bp)
     app.register_blueprint(diagnoses.omim_bp)
     app.register_blueprint(institutes.overview)
     app.register_blueprint(managed_variants.managed_variants_bp)
