+++ scout/server/blueprints/institutes/templates/overview/institute_sidebar.html
@@ -60,7 +60,7 @@
       </a>
 
       <!-- Advanced phenotypes -->
-      <a href="{{ url_for('overview.advanced_phenotypes', institute_id=institute._id) }}" class="bg-dark list-group-item list-group-item-action flex-column align-items-start">
+      <a href="{{ url_for('phenomodels.advanced_phenotypes', institute_id=institute._id) }}" class="bg-dark list-group-item list-group-item-action flex-column align-items-start">
           <div class="d-flex w-100 justify-content-start align-items-center">
               <span class="fa fa-stethoscope me-3"></span>
               <span class="menu-collapsed">Phenomodels</span>
