+++ scout/server/blueprints/phenomodels/__init__.py
@@ -1,0 +1,1 @@
+from .views import phenomodels_bp
